datachain-backend/pyproject.toml
[tool.poetry]
name = "datachain-backend"
version = "0.1.0"
description = ""
authors = ["Charles Calapini"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
openai = "^1.3.6"
requests = "^2.31.0"
tenacity = "^8.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
